requirements.txt
python-telegram-bot==20.7
google-generativeai==0.3.2
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
requests==2.31.0
beautifulsoup4==4.12.2
duckduckgo-search==4.1.1
schedule==1.2.0
python-dotenv==1.0.0
aiofiles==23.2.1
